\name{GMWL}
\alias{GMWL}
\docType{data}
\title{
Global Meteoric Water Line
}
\description{
Parameters for the Global Meteoric Water Line fit to a global precipitation compilation in Bowen, et al. (2019).
}
\usage{data("GMWL")}
\format{
  The format is:
 num [1:6] slope, intercept, average d18O, sum of squares in d18O, root mean square error, number of samples
}
\source{
Bowen et al. (2019) Isotopes in the water cycle: Regional- to global-Scale patterns and applications. \emph{Annual Review of Earth and Planetary Sciences} \strong{47} 453--479. \doi{10.1146/annurev-earth-053018-060220}.
}
\examples{
data(GMWL)
}
\keyword{datasets}
