\name{isoWater}

\alias{isoWater}

\title{Discovery, Retrieval, and Analysis of Water Isotope Data}

\description{
wiDB_ functions provide interface to the public API of 
    the wiDB: build, check and submit queries, and receive and 
    unpack responses. Data analysis functions support Bayesian 
    inference of the source and source isotope composition of water 
    samples that may have experienced evaporation. Algorithms 
    adapted from Bowen et al. (2018) \href{https://doi.org/10.1007/s00442-018-4192-5}{doi:10.1007/s00442-018-4192-5}.
}

\author{
\strong{Maintainer}: Gabriel Bowen \email{gabe.bowen@utah.edu}  
\strong{Authors}: Gabriel Bowen
}